Move verifySignature to dedicated extension file
