+++ Sources/WebAuthn/Ceremonies/Registration/Formats/PackedAttestation.swift
@@ -15,8 +15,6 @@
 import Foundation
 import SwiftCBOR
 import X509
-import Crypto
-import _CryptoExtras
 
 struct PackedAttestation {
     enum PackedAttestationError: Error {
@@ -77,7 +75,7 @@ struct PackedAttestation {
             guard case .validCertificate = verifierResult else {
                 throw PackedAttestationError.invalidLeafCertificate
             }
-
+            
             // 2. Verify signature
             // 2.1 Determine key type (with new Swift ASN.1/ Certificates library)
             // 2.2 Create corresponding public key object (EC2PublicKey/RSAPublicKey/OKPPublicKey)
@@ -94,41 +92,6 @@ struct PackedAttestation {
             }
 
             try credentialPublicKey.verify(signature: Data(sig), data: verificationData)
-        }
-    }
-}
-
-
-extension Certificate.PublicKey {
-    func verifySignature(_ signature: Data, algorithm: Certificate.SignatureAlgorithm, data: Data) throws -> Bool {
-        switch algorithm {
-
-        case .ecdsaWithSHA256:
-            guard let key = P256.Signing.PublicKey(self) else {
-                return false
-            }
-            let signature = try P256.Signing.ECDSASignature(derRepresentation: signature)
-            return key.isValidSignature(signature, for: data)
-        case .ecdsaWithSHA384:
-            guard let key = P384.Signing.PublicKey(self) else {
-                return false
-            }
-            let signature = try P384.Signing.ECDSASignature(derRepresentation: signature)
-            return key.isValidSignature(signature, for: data)
-        case .ecdsaWithSHA512:
-            guard let key = P521.Signing.PublicKey(self) else {
-                return false
-            }
-            let signature = try P521.Signing.ECDSASignature(derRepresentation: signature)
-            return key.isValidSignature(signature, for: data)
-        case .sha1WithRSAEncryption, .sha256WithRSAEncryption, .sha384WithRSAEncryption, .sha512WithRSAEncryption:
-            guard let key = _RSA.Signing.PublicKey(self) else {
-                return false
-            }
-            let signature = _RSA.Signing.RSASignature(rawRepresentation: signature)
-            return key.isValidSignature(signature, for: data)
-        default: // Should we return more explicit info (signature alg not supported) in that case?
-            return false
         }
     }
 }
